added the checkbox in the demo
