+++ example/lib/screens/demo/comic/widgets/check_box.demo.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class CheckBoxDemo extends StatefulWidget {
+  const CheckBoxDemo({
+    super.key,
+  });
+
+  @override
+  State<CheckBoxDemo> createState() => _CheckBoxDemoState();
+}
+
+class _CheckBoxDemoState extends State<CheckBoxDemo> {
+  bool value = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: () {
+        setState(() {
+          value = !value;
+        });
+      },
+      child: Row(
+        children: [
+          Checkbox(
+            value: value,
+            onChanged: null,
+          ),
+          const Expanded(
+            child: Text("Do you agree in the secret?"),
+          ),
+        ],
+      ),
+    );
+  }
+}
